Integrate with cv namespace, change directory structure
