+++ modules/dynamicfusion/doc/dynamicfusion.bib
@@ -1,0 +1,23 @@
+@InProceedings{Newcombe_2015_CVPR,
+author = {Newcombe, Richard A. and Fox, Dieter and Seitz, Steven M.},
+title = {DynamicFusion: Reconstruction and Tracking of Non-Rigid Scenes in Real-Time},
+booktitle = {The IEEE Conference on Computer Vision and Pattern Recognition (CVPR)},
+month = {June},
+year = {2015}
+}
+
+@inbook{innmann2016volume,
+author = "Innmann, Matthias and Zollh{\"o}fer, Michael and Nie{\ss}ner, Matthias and Theobalt, Christian
+         and Stamminger, Marc",
+editor = "Leibe, Bastian and Matas, Jiri and Sebe, Nicu and Welling, Max",
+title = "VolumeDeform: Real-Time Volumetric Non-rigid Reconstruction",
+bookTitle = "Computer Vision -- ECCV 2016: 14th European Conference, Amsterdam, The Netherlands,
+            October 11-14, 2016, Proceedings, Part VIII",
+year = "2016",
+publisher = "Springer International Publishing",
+address = "Cham",
+pages = "362--379",
+isbn = "978-3-319-46484-8",
+doi = "10.1007/978-3-319-46484-8_22",
+url = "http://dx.doi.org/10.1007/978-3-319-46484-8_22"
+}
