+++ modules/dynamicfusion/include/opencv2/dynamicfusion.hpp
@@ -1,0 +1,5 @@
+#ifndef OPENCV_DYNAMICFUSION_HPP
+#define OPENCV_DYNAMICFUSION_HPP
+#include <opencv2/kfusion/kinfu.hpp>
+#include <opencv2/kfusion/cuda/tsdf_volume.hpp>
+#endif //OPENCV_DYNAMICFUSION_HPP
