+++ modules/dynamicfusion/src/core.cpp
@@ -1,10 +1,10 @@
-#include <kfusion/kinfu.hpp>
+#include <opencv2/kfusion/kinfu.hpp>
 #include "safe_call.hpp"
 
 #include <cuda.h>
 #include <cstdio>
 #include <iostream>
-
+using namespace cv;
 int kf::cuda::getCudaEnabledDeviceCount()
 {
     int count;
