+++ modules/dynamicfusion/src/precomp.cpp
@@ -1,6 +1,6 @@
 #include "precomp.hpp"
 #include "internal.hpp"
-
+using namespace cv;
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 /// Kinfu/types implementation
